First R Course (IST 687)
This was the first R class (IST 687), and it was an eye opening experience. I started very green, but grew into R fairly rapidly. Eventually culminating in work with NOAA's API, and utilizing loops to pull data in a very efficient manner. The HW_6 file was also an important first step to understanding the GGLOT library. With this, I expanded my horizon into advanced visualizations techniques in leaflet, TMAP, and Plotly within the next few months.