Text Mining Tweets Project
An interesting way to use unstructured data to tell a story. Attached shows the sourcing of tweets, and then from there a way to predict word usage that may be indicative of a presidential win. With the Election in Two days can this be a good way to predict elections going forward? Only time will tell