Text Mining and NLP in Python
Analyzing tweets to predict a presidential election. What was interesting, was a lot of the word usage Biden was using (especially in the primary) was much closer to Clinton's tweets in the 2016 election, than an Obama or Trump winning campaign. Will that be problem getting voters out to the polls? Only time will tell...